+++ lib/plugins/aws/package/compile/events/alexaSmartHome.js
