+++ lib/plugins/index.js
@@ -43,7 +43,7 @@ module.exports = [
   require('./aws/package/compile/events/msk/index.js'),
   require('./aws/package/compile/events/alb/index.js'),
   require('./aws/package/compile/events/alexaSkill.js'),
-  require('./aws/package/compile/events/alexaSmartHome/index.js'),
+  require('./aws/package/compile/events/alexaSmartHome.js'),
   require('./aws/package/compile/events/iot/index.js'),
   require('./aws/package/compile/events/cloudWatchEvent/index.js'),
   require('./aws/package/compile/events/cloudWatchLog/index.js'),
