+++ test/unit/lib/plugins/aws/package/compile/events/alexaSmartHome.test.js
@@ -1,9 +1,9 @@
 'use strict';
 
 const expect = require('chai').expect;
-const AwsProvider = require('../../../../../../../../../lib/plugins/aws/provider/awsProvider');
-const AwsCompileAlexaSmartHomeEvents = require('../../../../../../../../../lib/plugins/aws/package/compile/events/alexaSmartHome/index');
-const Serverless = require('../../../../../../../../../lib/Serverless');
+const AwsProvider = require('../../../../../../../../lib/plugins/aws/provider/awsProvider');
+const AwsCompileAlexaSmartHomeEvents = require('../../../../../../../../lib/plugins/aws/package/compile/events/alexaSmartHome');
+const Serverless = require('../../../../../../../../lib/Serverless');
 
 describe('AwsCompileAlexaSmartHomeEvents', () => {
   let serverless;
